+++ src/qmflows/parsers/xyzParser.py
@@ -1,15 +1,15 @@
-__author__ = "Felipe Zapata"
+"""XYZ file format readers."""
 
 __all__ = ['parse_string_xyz', 'readXYZ', 'manyXYZ', "string_to_plams_Molecule"]
 
-# ===================> Standard libraries and third-party <====================
-from scm.plams import (Atom, Molecule)
-from pyparsing import (
-    alphas, Group, LineEnd, OneOrMore, restOfLine, Suppress, Word)
+from pyparsing import (Group, LineEnd, OneOrMore, Suppress, Word, alphas,
+                       restOfLine)
+from scm.plams import Atom, Molecule
 
 from qmflows.common import AtomXYZ
-from qmflows.parsers.parser import (floatNumber, natural)
+from qmflows.parsers.parser import floatNumber, natural
 from qmflows.utils import zipWith
+
 # =============================================================================
 
 header = natural + LineEnd() + restOfLine
@@ -63,16 +63,14 @@ def manyXYZ(pathXYZ):
 
 
 def createAtoms(xs):
-    """
-    Create an AtomXYZ tuple from a string.
-    """
+    """Create an AtomXYZ tuple from a string."""
     ls = [a.label.lower() for a in xs]
     rs = [list(map(float, a.xyz)) for a in xs]
     return zipWith(AtomXYZ)(ls)(rs)
 
 
 def tuplesXYZ_to_plams(xs):
-    """ Transform a list of namedTuples to a Plams molecule """
+    """Transform a list of namedTuples to a Plams molecule."""
     plams_mol = Molecule()
     for at in xs:
         symb = at.symbol
@@ -83,5 +81,5 @@ def tuplesXYZ_to_plams(xs):
 
 
 def string_to_plams_Molecule(xs):
-    """Convert a molecule stored in a string to a plams Molecule"""
+    """Convert a molecule stored in a string to a plams Molecule."""
     return tuplesXYZ_to_plams(parse_string_xyz(xs))